+++ covalent_slurm_plugin/slurm.py
@@ -21,18 +21,24 @@
 """Slurm executor plugin for the Covalent dispatcher."""
 
 import os
+import re
 import shutil
 import subprocess
 import tempfile
-import re
 import time
 from copy import deepcopy
-from typing import Any, Dict, List, Tuple
+from multiprocessing import Queue as MPQ
+from typing import Any, Dict, List, Union
 
 import cloudpickle as pickle
+from covalent._results_manager.result import Result
+from covalent._shared_files import logger
 from covalent._shared_files.util_classes import DispatchInfo
 from covalent._workflow.transport import TransportableObject
 from covalent.executor import BaseExecutor
+
+app_log = logger.app_log
+log_stack_info = logger.log_stack_info
 
 _EXECUTOR_PLUGIN_DEFAULTS = {
     "username": "",
@@ -86,28 +92,52 @@ class SlurmExecutor(BaseExecutor):
         function: TransportableObject,
         args: List,
         kwargs: Dict,
+        info_queue: MPQ,
+        task_id: int,
         dispatch_id: str,
         results_dir: str,
-        node_id: int = -1,
-    ) -> Tuple[Any, str, str]:
+    ) -> Any:
+        """
+        Executes the input function and returns the result.
+
+        Args:
+            function: The input python function which will be executed and whose result
+                      is ultimately returned by this function.
+            args: List of positional arguments to be used by the function.
+            kwargs: Dictionary of keyword arguments to be used by the function.
+            info_queue: A multiprocessing Queue object used for shared variables across
+                processes. Information about, eg, status, can be stored here.
+            task_id: The ID of this task in the bigger workflow graph.
+            dispatch_id: The unique identifier of the external lattice process which is
+                         calling this function.
+            results_dir: The location of the results directory.
+
+        Returns:
+            output: The result of the executed function.
+        """
 
         dispatch_info = DispatchInfo(dispatch_id)
-        result_filename = f"result-{dispatch_id}-{node_id}.pkl"
-        slurm_filename = f"slurm-{dispatch_id}-{node_id}.sh"
+        result_filename = f"result-{dispatch_id}-{task_id}.pkl"
+        slurm_filename = f"slurm-{dispatch_id}-{task_id}.sh"
         task_results_dir = os.path.join(results_dir, dispatch_id)
 
         if "output" not in self.options:
             self.options["output"] = os.path.join(
-                self.remote_workdir, f"stdout-{dispatch_id}-{node_id}.log"
+                self.remote_workdir, f"stdout-{dispatch_id}-{task_id}.log"
             )
         if "error" not in self.options:
             self.options["error"] = os.path.join(
-                self.remote_workdir, f"stderr-{dispatch_id}-{node_id}.log"
+                self.remote_workdir, f"stderr-{dispatch_id}-{task_id}.log"
             )
 
+        result = None
+        exception = None
+
+        info_dict = {"STATUS": Result.RUNNING}
+        info_queue.put_nowait(info_dict)
+
         with self.get_dispatch_context(dispatch_info), tempfile.NamedTemporaryFile(
-            dir=self.cache_dir, delete=False
-        ) as f, tempfile.NamedTemporaryFile(dir=self.cache_dir, mode="w", delete=False) as g:
+            dir=self.cache_dir) as f, tempfile.NamedTemporaryFile(dir=self.cache_dir, mode="w") as g:
 
             # Write the serialized function to file
             pickle.dump(function, f)
@@ -135,8 +165,9 @@ class SlurmExecutor(BaseExecutor):
             )
 
             # Copy the function to the remote filesystem
-            func_filename = f"func-{dispatch_id}-{node_id}.pkl"
+            func_filename = f"func-{dispatch_id}-{task_id}.pkl"
             remote_func_filename = os.path.join(self.remote_workdir, func_filename)
+
             subprocess.run(
                 [
                     "rsync",
@@ -150,11 +181,13 @@ class SlurmExecutor(BaseExecutor):
                 capture_output=True,
             )
 
+            func_py_version = ".".join(function.python_version.split(".")[:2])
+
             # Format the SLURM submit script
             slurm_submit_script = self._format_submit_script(
                 func_filename,
                 result_filename,
-                function.python_version,
+                func_py_version,
                 args,
                 kwargs,
             )
@@ -163,7 +196,7 @@ class SlurmExecutor(BaseExecutor):
 
             # Copy the script to the local results directory
             local_slurm_filename = os.path.join(task_results_dir, "slurm.sh")
-            shutil.copyfile(g.name, local_slurm_filename)
+            #shutil.copyfile(g.name, local_slurm_filename)
 
             # Copy the script to the remote filesystem
             remote_slurm_filename = os.path.join(self.remote_workdir, slurm_filename)
@@ -201,16 +234,22 @@ class SlurmExecutor(BaseExecutor):
                 ],
                 capture_output=True,
             )
-
-			# proc.stdout is `Submitted batch job <jobid>`
             if proc.returncode == 0:
-                slurm_job_id = int(re.findall('[0-9]+', proc.stdout.decode("utf-8"))[0])
+                slurm_job_id = int(re.findall("[0-9]+", proc.stdout.decode("utf-8"))[0])
             else:
                 raise Exception(proc.stderr)
 
             self._poll_slurm(slurm_job_id)
 
-            return self._query_result(result_filename, task_results_dir)
+            result, stdout, stderr, exception = self._query_result(
+                result_filename, task_results_dir
+            )
+
+            info_dict = info_queue.get()
+            info_dict["STATUS"] = Result.FAILED if result is None else Result.COMPLETED
+            info_queue.put(info_dict)
+
+            return result, stdout, stderr, exception
 
     def _format_submit_script(
         self,
@@ -261,9 +300,9 @@ fi
         )
 
         slurm_python_version = """
-if [[ "{python_version}" != `python -V | awk '{{print $2}}'` ]] ; then
-  >&2 echo "Python version mismatch. Please install Python {python_version} in the compute "\
-  "environment."
+remote_py_version=$(python -c "print(__import__('sys').version_info[0])").$(python -c "print(__import__('sys').version_info[1])")
+if [[ "{python_version}" != $remote_py_version ]] ; then
+  >&2 echo "Python version mismatch. Please install Python {python_version} in the compute environment."
   exit 199
 fi
 """.format(
@@ -277,10 +316,16 @@ import cloudpickle as pickle
 with open("{func_filename}", "rb") as f:
     function = pickle.load(f).get_deserialized()
 
-result = function(*{args}, **{kwargs})
+result = None
+exception = None
+
+try:
+    result = function(*{args}, **{kwargs})
+except Exception as e:
+    exception = e
 
 with open("{result_filename}", "wb") as f:
-    pickle.dump(result, f)
+    pickle.dump((result, exception), f)
 EOF
 
 wait
@@ -295,15 +340,20 @@ wait
 
         return slurm_submit_script
 
-    def get_status(self, job_id: str) -> str:
+    def get_status(self, info_dict: dict) -> Union[Result, str]:
         """Query the status of a job previously submitted to Slurm.
 
         Args:
-            job_id: Slurm job ID.
+            info_dict: a dictionary containing all neccessary parameters needed to query the
+                status of the execution. Required keys in the dictionary are:
+                    A string mapping "job_id" to Slurm job ID.
 
         Returns:
             status: String describing the job status.
         """
+        job_id = info_dict.get("job_id", None)
+        if job_id is None:
+            return Result.NEW_OBJ
 
         proc = subprocess.run(
             [
@@ -338,7 +388,8 @@ wait
         """
 
         # Poll status every `poll_freq` seconds
-        status = self.get_status(str(job_id))
+        status = self.get_status({"job_id": str(job_id)})
+        #status = self.get_status(str(job_id))
         while (
             "PENDING" in status
             or "RUNNING" in status
@@ -346,7 +397,8 @@ wait
             or "CONFIGURING" in status
         ):
             time.sleep(self.poll_freq)
-            status = self.get_status(str(job_id))
+            status = self.get_status({"job_id": str(job_id)})
+            #status = self.get_status(str(job_id))
 
         if "COMPLETED" not in status:
             raise Exception("Job failed with status:\n", status)
@@ -392,7 +444,7 @@ wait
                 f"ssh -i {self.ssh_key_file} -o StrictHostKeyChecking=no "
                 "-o UserKnownHostsFile=/dev/null -o LogLevel=ERROR",
                 f"{self.username}@{self.address}:{remote_result_filename}",
-                task_results_dir,
+                task_results_dir+"/",
             ],
             check=True,
             capture_output=True,
@@ -406,7 +458,7 @@ wait
                 f"ssh -i {self.ssh_key_file} -o StrictHostKeyChecking=no "
                 "-o UserKnownHostsFile=/dev/null -o LogLevel=ERROR",
                 f"{self.username}@{self.address}:{self.options['output']}",
-                task_results_dir,
+                task_results_dir+"/",
             ],
             check=True,
             capture_output=True,
@@ -418,7 +470,7 @@ wait
                 f"ssh -i {self.ssh_key_file} -o StrictHostKeyChecking=no "
                 "-o UserKnownHostsFile=/dev/null -o LogLevel=ERROR",
                 f"{self.username}@{self.address}:{self.options['error']}",
-                task_results_dir,
+                task_results_dir+"/",
             ],
             check=True,
             capture_output=True,
@@ -426,7 +478,7 @@ wait
 
         local_result_filename = os.path.join(task_results_dir, result_filename)
         with open(local_result_filename, "rb") as f:
-            result = pickle.load(f)
+            result, exception = pickle.load(f)
         os.remove(local_result_filename)
 
         stdout_file = os.path.join(task_results_dir, os.path.basename(self.options["output"]))
@@ -439,14 +491,4 @@ wait
             stderr = f.read()
         os.remove(stderr_file)
 
-        return result, stdout, stderr
-
-    def cancel(self, job_id: int) -> None:
-        """Cancel a Slurm job.
-
-        Args:
-            job_id: Slurm job ID.
-
-        Returns:
-            None
-        """
+        return result, stdout, stderr, exception